+++ src/app/[locale]/dashboard/@savedData/default.tsx
@@ -1,7 +1,7 @@
 import TabSelector from './TabsSelector';
 import IntlWrapperForAsync from '@/components/IntlWrapperForAsync/IntlWrapperForAsync';
-import History from '../[id]/history/History';
-import SavedBoard from '../[id]/savedBoards/SavedBoard';
+import History from './_components/history/History';
+import SavedBoard from './_components/savedBoards/SavedBoard';
 
 export default async function Default() {
   return (
